revised the method to determine the amount of moisture needed to form clouds in a grid volume.  This is now computed in the routine lq3_driver1a.f and passed into the subsequent routines variational.f and then the functions. Note also that a new version of the makefile includes the new routine test_cloud.f  Test_cloud.f is part of this addition and will be added to CVS during another action.
This new module is needed for the changes cited above.  It uses an
approximation for the amount of relative humidity change in a grid volume.
This change or variablity is 20 times more sensative to temperature than
specific humidity (at least on a percent change level).  This routine now
computes the likely variability in a grid volume of RH and from this determines
the amount of additional moisture to enable cloud formation somewhere in the
volume.  In the new lq3 scheme, this is the amount of moisture that is added
when the volume is adjusted for cloud saturation.  This will be a much more
conservative adjustment than is currently done.  The complete saturation method
has been shown to overcorrect for cloud moisture when benchmarked against GPS
data.

The new scheme is auto scaled.  As the grid volumme goes to zero size (a point)
the variability or uncertainty of RH goes to the error in measurement of
temperature and specific humidity.