- filter_filenames and get_dir_length: mods for Taiwan CWB new gfs and tfs model background grids; ie., the model bkgd names start with gf and tf, thus we must not filter these.
- laps_io.f and read_namelist.f: mods to subroutine get_background_info
accommodating new background.nl namelist variable "luse_bkgd_sfc" (subroutine
get_background_info used in routines in file laps_io.f.
New namelist variable allows model background 2m sfc variables to used in
lgb calculations.
laps_io.f has new constructs designed to determine which type of background
is used (ie., ruc or avn or eta and so forth) so that eventually backgrounds
for laps for all models are made and put into subdirectories.  This code runs
but is not functional for the eventual model background verification work that
it is intended for.