Improved version of get_newest_file function
The prior form of the function failed to provide the return of the
newest file in the interval selected due to a coding error.  Prior
to this fix the oldest file was returned (in the allowed interval)
now the "newest" or most recent file in the allowed interval is
returned.  - Dan Birkenheuer